pyproject.toml
[project]
name = "goit-pythonweb-hw-06"
version = "0.1.0"
description = ""
authors = [
    {name = "sanyanut",email = "alexander.sorochinskiy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "faker (>=40.15.0,<41.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
